Document-Summarizer/requirements/requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
fastapi==0.63.0
filelock==3.0.12
Flask==1.1.2
h11==0.12.0
idna==2.10
importlib-metadata==3.4.0
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.1
lxml==4.6.2
MarkupSafe==1.1.1
numpy==1.20.1
packaging==20.9
Pillow==8.1.0
pydantic==1.7.3
pyparsing==2.4.7
python-docx==0.8.10
python-multipart==0.0.5
regex==2020.11.13
requests==2.25.1
sacremoses==0.0.43
sentencepiece==0.1.95
six==1.15.0
starlette==0.13.6
tokenizers==0.10.1
torch==1.7.1
torchaudio==0.7.2
torchvision==0.8.2
tqdm==4.56.2
transformers==4.3.2
typing-extensions==3.7.4.3
urllib3==1.26.3
uvicorn==0.13.3
Werkzeug==1.0.1
zipp==3.4.0
